fixed bug where logout didn't work if your session had expired or you'd never logged in; fixed redirect in GuardianActivation if no email

+++ modules/users/app/controllers/users/App.scala
@@ -15,9 +15,11 @@ import models.users.{ Activation, QUser, User, Visit }
 import scala.xml.NodeSeq
 import scalatags._
 import play.api.libs.json.Json._
+import javax.jdo.JDOHelper
+import javax.jdo.ObjectState
 
 @Singleton
-class App @Inject()(implicit config: Config) extends Controller with UsesDataStore {  
+class App @Inject() (implicit config: Config) extends Controller with UsesDataStore {
   object LoginForm extends Form {
     val username = new TextField("username")
     val password = new PasswordField("password")
@@ -25,10 +27,10 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
     def fields = List(username, password)
 
     override def validate(vb: ValidBinding): ValidationError = {
-        User.authenticate(vb.valueOf(username), vb.valueOf(password)) match {
-          case None => ValidationError("Incorrect username or password.")
-          case Some(user) => ValidationError(Nil)
-        }
+      User.authenticate(vb.valueOf(username), vb.valueOf(password)) match {
+        case None => ValidationError("Incorrect username or password.")
+        case Some(user) => ValidationError(Nil)
+      }
     }
   }
 
@@ -43,38 +45,38 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
   }
 
   def loginP() = VisitAction { implicit request =>
-      Binding(LoginForm, request) match {
-        case ib: InvalidBinding => Ok(templates.users.Login(ib))
-        case vb: ValidBinding => {
-          // set the session user
-          request.visit.user = User.getByUsername(vb.valueOf(LoginForm.username))
-          // set the session role
-          request.visit.user.map(_.roles).getOrElse(Nil) match {
-            // no roles attached to this user
-            case Nil => Redirect(controllers.users.routes.App.login()).flashing("message" -> "That user has no active roles.")
-            // there's at least one
-            case role :: rest => {
-              // TODO: should we have a default role?
-              // set the first one
-             dataStore.execute { pm =>
+    Binding(LoginForm, request) match {
+      case ib: InvalidBinding => Ok(templates.users.Login(ib))
+      case vb: ValidBinding => {
+        // set the session user
+        request.visit.user = User.getByUsername(vb.valueOf(LoginForm.username))
+        // set the session role
+        request.visit.user.map(_.roles).getOrElse(Nil) match {
+          // no roles attached to this user
+          case Nil => Redirect(controllers.users.routes.App.login()).flashing("message" -> "That user has no active roles.")
+          // there's at least one
+          case role :: rest => {
+            // TODO: should we have a default role?
+            // set the first one
+            dataStore.execute { pm =>
               request.visit.role = Some(role)
               request.visit.permissions = role.permissions
               request.visit.updateMenu()
               pm.makePersistent(request.visit)
-             }
-              rest match {
-                // there was only one
-                case Nil => {
-                  if (request.visit.redirectUrl.isDefined) Redirect(request.visit.redirectUrl.get)
-                  else Redirect(config.defaultCall).flashing("message" -> "You have successfully logged in.")
-                }
-                // multiple roles
-                case _ => Redirect(controllers.users.routes.App.chooseRole()).flashing("message" -> "Choose which role to use.")
+            }
+            rest match {
+              // there was only one
+              case Nil => {
+                if (request.visit.redirectUrl.isDefined) Redirect(request.visit.redirectUrl.get)
+                else Redirect(config.defaultCall).flashing("message" -> "You have successfully logged in.")
               }
+              // multiple roles
+              case _ => Redirect(controllers.users.routes.App.chooseRole()).flashing("message" -> "Choose which role to use.")
             }
           }
         }
       }
+    }
   }
 
   class ChooseRoleForm(visit: Visit) extends Form {
@@ -94,18 +96,18 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
   }
 
   def chooseRoleP() = Authenticated { implicit req =>
-      val form = new ChooseRoleForm(req.visit)
-      Binding(form, req) match {
-        case ib: InvalidBinding => Ok(templates.users.ChooseRole(ib))
-        case vb: ValidBinding => {
-          req.visit.role = Some(vb.valueOf(form.role))
-          req.visit.permissions = req.visit.role.map(_.permissions).getOrElse(Set())
-          req.visit.updateMenu
-          dataStore.pm.makePersistent(req.visit)
+    val form = new ChooseRoleForm(req.visit)
+    Binding(form, req) match {
+      case ib: InvalidBinding => Ok(templates.users.ChooseRole(ib))
+      case vb: ValidBinding => {
+        req.visit.role = Some(vb.valueOf(form.role))
+        req.visit.permissions = req.visit.role.map(_.permissions).getOrElse(Set())
+        req.visit.updateMenu
+        dataStore.pm.makePersistent(req.visit)
         Redirect(req.visit.redirectUrl.getOrElse(config.defaultCall)).flashing("message" -> "You have successfully logged in.")
-        }
       }
     }
+  }
 
   /**
    * regex: logout
@@ -115,7 +117,9 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
 
   def logout = VisitAction { implicit req =>
     dataStore.execute { pm =>
-      pm.deletePersistent(req.visit)
+      if (JDOHelper.isPersistent(req.visit)) {
+        pm.deletePersistent(req.visit)
+      }
     }
     Redirect(config.defaultCall).flashing("message" -> "You have been logged out.")
   }
@@ -161,12 +165,11 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
     Ok(templates.users.ChangeSettings(Binding(pwForm), Binding(ChangeTheme)))
   }*/
 
-  
-  def changePassword() = Authenticated { implicit req => 
-    val form = new ChangePasswordForm(req.role.user)  
+  def changePassword() = Authenticated { implicit req =>
+    val form = new ChangePasswordForm(req.role.user)
     Ok(templates.users.ChangePassword(Binding(form)))
   }
-  
+
   def changePasswordP() = Authenticated { implicit req =>
     val user = req.role.user
     val form = new ChangePasswordForm(user)
@@ -178,12 +181,12 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
       }
     }
   }
-  
+
   class ActivationForm(uuid: String) extends Form with UsesDataStore {
     val username = new TextField("username")
     val newPassword = new PasswordField("newPassword")
     val verifyNewPassword = new PasswordField("verifyPassword")
-    
+
     val defaultError = ValidationError("The username and activation URL do not match. Check and re-try.")
 
     override def validate(vb: ValidBinding): ValidationError = Activation.getByUuid(uuid) match {
@@ -199,20 +202,24 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
         } else ValidationError(Nil)
       }
     }
-    
+
     def fields = List(username, newPassword, verifyNewPassword)
   }
-  
+
   def activate(uuid: String) = VisitAction { implicit req =>
-    Ok(templates.users.Activate(Binding(new ActivationForm(uuid))))
+    Activation.getByUuid(uuid) match {
+      case Some(activation) => Ok(templates.users.Activate(Binding(new ActivationForm(uuid))))
+      case None => Redirect(controllers.users.routes.App.login).flashing(
+        "message" -> "This activation has already been used. Log in with the username provided and the password you chose.")
+    }
   }
-  
+
   def activateP(uuid: String) = VisitAction { implicit req =>
     val form = new ActivationForm(uuid)
     Binding(form, req) match {
       case ib: InvalidBinding => Ok(templates.users.Activate(ib))
       case vb: ValidBinding => {
-        dataStore.execute( pm => {
+        dataStore.execute(pm => {
           val activation = Activation.getByUuid(uuid).get
           activation.user.password = vb.valueOf(form.newPassword)
           pm.makePersistent(activation.user)
@@ -222,7 +229,7 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
       }
     }
   }
-  
+
   object ChangeOtherPasswordForm extends Form {
     val username = new TextField("username") {
       override def validators: List[Validator[String]] = {
@@ -242,21 +249,19 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
         ValidationError("New password and verify password must match.")
       } else ValidationError(Nil)
     }
-      
-    val fields = List(username, newPassword, verifyNewPassword)      
+
+    val fields = List(username, newPassword, verifyNewPassword)
   }
-  
-  
+
   def activeUserList() = PermissionRequired(User.Permissions.ListAll) { implicit req =>
     val users = User.activeUsers.map(u => toJsFieldJsValueWrapper(Map("username" -> u.username, "formalName" -> u.formalName, "email" -> u.email.getOrElse(""))))
     Ok(obj("users" -> arr(users: _*)))
   }
-  
-  
-  def changeOtherPassword() = PermissionRequired(User.Permissions.ChangePassword) { implicit req => 
-    Ok(templates.users.ChangeOtherPassword(Binding(ChangeOtherPasswordForm)))  
+
+  def changeOtherPassword() = PermissionRequired(User.Permissions.ChangePassword) { implicit req =>
+    Ok(templates.users.ChangeOtherPassword(Binding(ChangeOtherPasswordForm)))
   }
-  
+
   def changeOtherPasswordP() = PermissionRequired(User.Permissions.ChangePassword) { implicit req =>
     val form = ChangeOtherPasswordForm
     Binding(ChangeOtherPasswordForm, req) match {
@@ -268,7 +273,7 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
       }
     }
   }
-  
+
   /**
    * regex: /listUsers
    *
@@ -278,19 +283,19 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
     val cand = QUser.candidate
     Ok(templates.users.ListUsers(dataStore.pm.query[User].orderBy(cand.last.asc, cand.first.asc).executeList()))
   }
-  
+
   class ChooseUserForm extends Form {
     val cand = QUser.candidate()
     val allUsers = dataStore.pm.query[User].filter(cand.isActive.eq(true)).orderBy(cand.last.asc, cand.first.asc).executeList()
     val user = new ChoiceField("user", allUsers.map(u => (u.formalName, u.id)))
-    
+
     def fields = List(user)
   }
-  
+
   def chooseUserToEdit() = PermissionRequired(User.Permissions.Manage) { implicit req =>
-    Ok(templates.users.ChooseUser(Binding(new ChooseUserForm)))  
+    Ok(templates.users.ChooseUser(Binding(new ChooseUserForm)))
   }
-  
+
   def chooseUserToEditP() = PermissionRequired(User.Permissions.Manage) { implicit req =>
     val form = new ChooseUserForm()
     Binding(form, req) match {
@@ -301,7 +306,7 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
       }
     }
   }
-  
+
   class UpdateUserForm(user: User) extends Form {
     // TODO: make sure this is unique
     val username = new TextField("username") { override def initialVal = Some(user.username) }
@@ -310,10 +315,10 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
     val middleName = new TextFieldOptional("middleName") { override def initialVal = Some(user.middle) }
     val preferredName = new TextFieldOptional("preferredName") { override def initialVal = Some(user.preferred) }
     val email = new TextFieldOptional("email") { override def initialVal = Some(user.email) }
-  
+
     def fields = List(username, lastName, firstName, middleName, preferredName, email)
   }
-  
+
   def editUser(userId: Long) = PermissionRequired(User.Permissions.Manage) { implicit req =>
     User.getById(userId) match {
       case None => NotFound("No user with the given id.")
@@ -323,7 +328,7 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
       }
     }
   }
-  
+
   def editUserP(userId: Long) = PermissionRequired(User.Permissions.Manage) { implicit req =>
     User.getById(userId) match {
       case None => NotFound("No user with the given id.")
@@ -343,7 +348,6 @@ class App @Inject()(implicit config: Config) extends Controller with UsesDataSto
         }
       }
     }
-    
-  
+
   }
 }
